requirements.txt

# server/requirements.txt (Main requirements file)
# KisanSathi AI Backend - Main Requirements
# Install with: pip install -r requirements.txt

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing and analysis
pandas==2.1.3
numpy==1.25.2
requests==2.31.0

# Date/time handling
python-dateutil==2.8.2

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# HTTP client improvements
httpx==0.25.2
aiohttp==3.9.1

# Caching (in-memory for now)
cachetools==5.3.2

# Optional: If you want to use your existing LangGraph agent
# Uncomment these lines if you're integrating your existing mandi agent
# langgraph==0.0.62
# langchain==0.0.352
# langchain-core==0.1.0

# Development dependencies (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0